added error handling in comment usecase

+++ business/comments/usecase.go
@@ -3,6 +3,7 @@ package comments
 import (
 	"context"
 	"disspace/helpers/messages"
+	"strings"
 	"time"
 )
 
@@ -19,6 +20,10 @@ func NewCommentUseCase(commentRepository Repository, timeout time.Duration) UseC
 }
 
 func (useCase *CommentUseCase) Create(ctx context.Context, commentDomain *Domain, id string) (Domain, error) {
+	if strings.TrimSpace(commentDomain.Text) == "" {
+		return Domain{}, messages.ErrTextCannotBeEmpty
+	}
+
 	result, err := useCase.commentRepo.Create(ctx, commentDomain, id)
 	if err != nil {
 		return Domain{}, err
@@ -27,6 +32,14 @@ func (useCase *CommentUseCase) Create(ctx context.Context, commentDomain *Domain
 }
 
 func (useCase *CommentUseCase) Delete(ctx context.Context, id string, threadId string) error {
+	if strings.TrimSpace(id) == "" {
+		return messages.ErrInvalidCommentID
+	}
+
+	if strings.TrimSpace(threadId) == "" {
+		return messages.ErrInvalidThreadID
+	}
+
 	err := useCase.commentRepo.Delete(ctx, id, threadId)
 	if err != nil {
 		if err == messages.ErrInvalidThreadID {
@@ -38,6 +51,10 @@ func (useCase *CommentUseCase) Delete(ctx context.Context, id string, threadId s
 }
 
 func (useCase *CommentUseCase) Search(ctx context.Context, q string, sort string) ([]Domain, error) {
+	if sort != "" && sort != "created" && sort != "num_votes" && sort != "num_comments" {
+		return []Domain{}, messages.ErrInvalidQueryParam
+	}
+
 	result, err := useCase.commentRepo.Search(ctx, q, sort)
 	if err != nil {
 		return []Domain{}, err
@@ -46,6 +63,10 @@ func (useCase *CommentUseCase) Search(ctx context.Context, q string, sort string
 }
 
 func (useCase *CommentUseCase) GetByID(ctx context.Context, id string) (Domain, error) {
+	if strings.TrimSpace(id) == "" {
+		return Domain{}, messages.ErrInvalidCommentID
+	}
+
 	result, err := useCase.commentRepo.GetByID(ctx, id)
 	if err != nil {
 		return Domain{}, messages.ErrDataNotFound
@@ -54,6 +75,10 @@ func (useCase *CommentUseCase) GetByID(ctx context.Context, id string) (Domain, 
 }
 
 func (useCase *CommentUseCase) GetAllInThread(ctx context.Context, threadId string, parentId string, option string) ([]Domain, error) {
+	if strings.TrimSpace(threadId) == "" || strings.TrimSpace(parentId) == "" {
+		return []Domain{}, messages.ErrInvalidThreadOrParent
+	}
+	
 	result, err := useCase.commentRepo.GetAllInThread(ctx, threadId, parentId, option)
 	if err != nil {
 		return []Domain{}, err
